pyproject.toml
[project]
name = "linux-arctis-manager"
version = "2.0.0-dev"
description = "A replacement for SteelSeries GG software, to manage your Arctis device on Linux!"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "dbus-next>=0.2.3",
    "pulsectl>=24.12.0",
    "pyside6>=6.10.1",
    "pyudev>=0.24.4",
    "pyusb>=1.3.1",
    "ruamel-yaml>=0.19.1",
]

[project.scripts]
lam-daemon = "linux_arctis_manager.scripts.daemon:main"
lam-cli = "linux_arctis_manager.scripts.cli:main"
lam-gui = "linux_arctis_manager.scripts.gui:main"

[build-system]
requires = ["uv_build>=0.9.17,<0.10.0"]
build-backend = "uv_build"

[dependency-groups]
dev = [
    "autoflake>=2.3.1",
    "isort>=7.0.0",
    "pytest>=9.0.2",
]

[tool.pyright]
reportUnknownMemberType = "none"
